+++ wpml-compatibility-test-tools.php
@@ -3,12 +3,12 @@
 Plugin Name: WPML Compatibility Test Tools
 Plugin URI: http://wpml.org
 Description: Tools to test theme and plugin compatibility.
-Author: WPML Development Team
+Author: WPML Compatibility Team
 Author URI: http://wpml.org
 Version: 1.1.1
 */
 
-define('WPML_CTT_VERSION', '1.1.1');
+define('WPML_CTT_VERSION', '1.1.2');
 define('WPML_CTT_PATH', dirname(__FILE__));
 define('WPML_CTT_FOLDER', basename(WPML_CTT_PATH));
 define('WPML_CTT_MENU_SETTINGS_SLUG', WPML_CTT_FOLDER . '/menus/settings/settings.php');