+++ changelog.md
@@ -3,10 +3,15 @@
 ## Improvements
 
 * Added checkboxes instead of option list when selecting contexts for translation.
+* Added "Toggle All" for admin text checkboxes.
+* Some code style improvements.
+* Checkboxes are now automatically checked on radio button change.
 
 ## Bugfix
 
-* Fixed: Removed notice, checked strings inputs are remembered.
+* WP Options with 'autoload' set as 'no' were not listed on configuration generator screen.
+* Escaped special characters in scripts.js
+* Removed notice, checked strings inputs are remembered.
 
 # 1.1.1